ACPI: MADT: add in compliance tests for the MADT and subtables
This patch adds in a series of detailed ACPI specification compliance
tests for the MADT and most of the possible subtables (tests for the
GIC ITS subtable will be added in a separate patch due to changes needed
in FWTS header files).

The added tests for the MADT include verifying the versions of the table
being used and basic content checks.  The revision of the MADT is also
checked against the version of the FADT to make sure they correspond to
each other.  We also verify that each subtable used is actually defined
for this revision of the MADT and that for each of the possible subtable
types, their contents are reasonably correct.

While this patch is based on the original src/acpi/madt.c file, it has
been reorganized so that each subtable has a separate function in order
to keep things clearer.

Signed-off-by: Al Stone <al.stone@linaro.org>