ACPI: MADT: further update to GICC flag checks for 6.0
In a previous patch ("ACPI: MADT: update GICC flag checks for ACPI 6.0"),
I corrected the flag mask used in the if test, but failed to correct the
flag used in the fwts_failed() call that was part of the logic block.

This patch fixes the mask for fwts_failed() so the test and failure message
are consistent again.

Signed-off-by: Al Stone <al.stone@linaro.org>