Add mechanism to tell FWTS what architecture is being tested
In some cases, we will want to run FWTS on one architecture, but use it
test firmware tables that may be for another architecture.  For example,
we may want to test ACPI tables for arm64 machines on an x86 laptop.

Previous patches provided the fwts_architecture as a typedef, along with
some helper functions.  Here, we add in framework variables for the host
(the machine running FWTS) and the target (the arch of the tables being
tested) and then initialize them properly.

Further, we add the --arch=<name> parameter so that the user can specify
the target architecture.  By default, the target will be assumed to be
the same as the host.

Signed-off-by: Al Stone <al.stone@linaro.org>