ACPI RSDP: flesh out the tests to check for full spec compliance.
Add in several more tests to the RSDP code that check for spec details
that had not been checked for previously.  This then allows us to tag
the RSDP tests so that they will also be executed when running FWTS to
check ACPI specification compliance (i.e., using the --acpicompliance
parameter).

Signed-off-by: Al Stone <ahs3@redhat.com>