FADT: enable compiling on non-x86 architectures
Instead of disabling the compilation of all FADT tests on non-x86 machines,
enable as much as possible, even if the tests do not yet make sense.  In
subsequent patches, we will modify the tests to acknowledge the specifics
of a given target arch.  This does not change functionality on x86, but it
does allow us to start testing portions of the FADT on other architectures.

Signed-off-by: Al Stone <al.stone@linaro.org>