FADT: add in code to log basic info about the various FADT flag fields
Add in some log messages that report on the state of all of the flag
fields in the FADT: flags, itself, plus the IA-PC boot architecture
flags and the ARM boot architecture flags.  Having this info in the
log should make it more straightforward to understand test results later
on in the FADT tests.

Signed-off-by: Al Stone <al.stone@linaro.org>