ACPI FADT: add spec compliance tests
This is a major rewrite and restructure of the FADT tests, with the goal
of building up ACPI specification compliance tests.  The tests are now
set up so there is essentially a function for each field of the FADT to
be tested.  Where possible, values are checked for correctness, sometimes
based on the values of other fields, or in one case dependent on flags
set in the FACS.  Further, testing for reduced hardware mode has been
incorporated so that tests that are irrelevant in that mode are never
run.

Unfortunately, there are still a couple of TODOs in the code.  As soon
as I can figure out a reasonable way to test the fields in question, I
will fix them.  In some cases, this will require execution of the AML
interpreter to traverse the ACPI namespace, which may be non-trivial.

Signed-off-by: Al Stone <al.stone@linaro.org>