FADT: restructure test sequence around reduced hardware mode
Since a prior patch added a test of of the reduced hardware fields, we
can now resequence the tests so that we only test the fields we must.

If we are not in reduced hardware mode, we test everything.  If we are,
we ignore tests for many fields.  They are simply irrelevant, and if the
fields are non-zero, that would have been caught in the prior patch.

There are also several fields that really cannot be reliably tested; we
cannot gather enough info to know whether they are correct or not, or
the field is allowed to be any value that it can hold.  Simply log the
value for these fields, but only when in the proper mode.

Signed-off-by: Al Stone <al.stone@linaro.org>