FADT: extend and add PM address block compliance tests
There are several address fields for power management in the FADT:
	PM1A_EVT_BLK
	PM1B_EVT_BLK
	PM1A_CNT_BLK
	PM1B_CNT_BLK
	PM2_CNT_BLK
	PM_TMR_BLK
	PM1_EVT_LEN
	PM1_CNT_LEN
	PM2_CNT_LEN
	PM_TMR_LEN

The tests that existed before only touched a few of these, so extend
the tests so that now all of them are checked for proper values.  Most
of the tests are very similar, hence they are grouped together into this
one patch.

Signed-off-by: Al Stone <al.stone@linaro.org>