FADT: add in compliance tests for C2/C3 latency fields
Add in new compliance checks for the P_LVL2_LAT and P_LVL3_LAT fields.
Both of these require knowing the value of an x86 P_BLK; this in turn
requires examination of any Processor() declarations in the ACPI name
space, which in turn requires the initialization of the ACPI interpreter.

It is probable that these fields are seldom used; processor speeds and
feeds are typically controlled via very different mechanisms these days.
These tests are therefore for completeness sake and it may be difficult
to find ACPI tables using these fields.

Note that these tests allow us to remove commented out versions of
simpler tests of these fields.

Signed-off-by: Al Stone <al.stone@linaro.org>